[STF] Relocate plugin start to stf project
