+++ eclipse/build.properties
@@ -1,7 +1,5 @@
 source.. = src/,\
            ext-src/,\
-           test/framework/stf/src/,\
-           test/framework/stf/test/,\
            test/junit/
 
 output.. = bin/
